backend/requirements.txt
fastapi~=0.115.11
uvicorn~=0.34.0
pydantic~=2.10.6
sqlalchemy
alembic
python-jose~=3.4.0
passlib~=1.7.4
python-multipart
email-validator
python-dotenv
bcrypt
httpx
itsdangerous
aiofiles
pyjwt
authlib
beanie~=1.29.0
motor~=3.7.0
pydantic-settings~=2.8.1
pip~=25.0.1
attrs~=22.2.0
wheel~=0.45.1
cryptography~=39.0.2
filelock~=3.9.0
setuptools~=75.8.2
packaging~=23.0
trove-classifiers~=2023.3.9
Jinja2~=3.1.6
typing_extensions~=4.12.2
build~=0.10.0
